Use "FETCH FIRST n ROWS ONLY" instead of ROWNUM for LIMIT for Oracle
